First, we need to initialize the callback and numbers. And then we need to create a inner class with layout: view, and create a textView,
later, we need to create a init, which setOnClickListener in it. and apply the position in it.

and in main activity, we need to define a data class "NumberSize", later on, modify the callback and use lambda.

finally, in textview layout, we need to change the layout height in to wrap_content, so we can see the gap between each number.